Move descriptor, weights, and truths print to end of process configs instead of end of solvers, also added in bikflag
