+++ fitsnap3/fitsnap.py
@@ -62,16 +62,18 @@ class FitSnap:
         for i, configuration in enumerate(self.data):
             self.calculator.process_configs(configuration, i)
         del self.data
+        self.calculator.extras()
 
     @pt.single_timeit
     def perform_fit(self):
-        if self.fit is None:
+        if not config.args.perform_fit:
+            return
+        elif self.fit is None:
             self.solver.perform_fit()
         else:
             self.solver.fit = self.fit
         self.solver.fit_gather()
         self.solver.error_analysis()
-        self.solver.extras()
 
     @pt.single_timeit
     def write_output(self):
